feat: :sparkles: add use reducer init
