+++ src/main.jsx
@@ -3,8 +3,12 @@ import ReactDOM from 'react-dom/client'
 import { RouterProvider } from "react-router-dom";
 import Router from "./router";
 import "./main.scss"
+import { HotelProvider } from './context';
 ReactDOM.createRoot(document.getElementById('root')).render(
+
   <React.StrictMode>
-    <RouterProvider router={Router} />
+    <HotelProvider>
+      <RouterProvider router={Router} />
+    </HotelProvider>
   </React.StrictMode>,
 )
